backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.21.3
anyio==4.13.0
APScheduler==3.10.4
bcrypt==4.1.2
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
filelock==3.25.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.73.1
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.2
huggingface_hub==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
packaging==26.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.3.2
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.7
